YouView application user lookup
As signed application in YouView ecosystem have generated UID from a given
range 0x4000-07bff their credentials are not available in typical locations,
namely /etc/passwd and /etc/groups. D-BUS doesn't like it and refuses such
users as it fails to obtain the credentials.
To mitigate that problem this patch intercepts D-BUS user credentails
lookup in _dbus_user_database_lookup(). If the UID is in the range above
0x4000 (which is reserved by YouView) it reads GID and supplementary group
GID array from /proc/<pid>/status as D-BUS daemon actually cares about
numeric values. User name and home directory are not set.

Some functions required additional PID argument and wherever it makes sense
it is obtained from DBusConnection (provided it is authenticated). When it
is not possible to get PID or it does not make sense (e. g. parsing
configuration files) it is simply not set (DBUS_PID_UNSET) and if it hits
our lookup code (it shouldn't) then it will simply fail.

Additionally the lookup is extended to YouView specific passwd and groups
files (fgetpwent() and  fgetgrent()) so that new daemon users can be added
to YouView specific files (using reserved range above 0x7c00).